removed unused au import, testing branching work started on cli interface
