+++ security_scripts/controls/tests/ec2fake.sh
