+++ security_scripts/controls/tests/red.sh
