+++ security_scripts/information/bin/.gitignore
