+++ security_scripts/information/bin/defaults.cfg
@@ -1,0 +1,17 @@
+[DEFAULT]
+# usins in all tools, but trailscraper has hard-coded need for Scimma to be the default. 
+profile=default
+[TAG_REPORT]
+loglevel=NORMAL
+dbfile=:memory:
+[FIND_BY_CONTENT]
+vaultdir=~/.vault
+loglevel=INFO
+[DOWNLOAD]
+loglevel=INFO
+bucket=s3://scimma-processes/Scimma-event-trail
+accountid=585193511743
+vaultdir=~/.vault
+[S3_REPORT]
+profile=scimma-uiuc-aws-admin
+loglevel=NORMAL
