+++ security_scripts/information/bin/find_by_content.py
