+++ security_scripts/information/bin/vault.py
