+++ security_scripts/information/bin/vault/.gitignore
