+++ security_scripts/information/bin/vault/2020-07-10-event_history.json
