+++ security_scripts/information/lib/aws_utils.py
