+++ security_scripts/information/lib/measurements.py
