+++ security_scripts/information/lib/vanilla_utils.py
